zappy_network::network

Trait ServerEventHandler

pub trait ServerEventHandler {
    // Required methods
    fn on_client_connected(
        &mut self,
        client_id: usize,
        network: &mut NetworkServer,
    );
    fn on_client_disconnected(
        &mut self,
        client_id: usize,
        network: &mut NetworkServer,
    );
    fn on_message_received(
        &mut self,
        client_id: usize,
        msg: String,
        network: &mut NetworkServer,
    );
    fn on_tick(&mut self, network: &mut NetworkServer) -> bool;
    fn get_next_timeout(&self) -> Duration;
}

Trait implemented by the application layer to react to network events.

Required Methods

fn on_client_connected(&mut self, client_id: usize, network: &mut NetworkServer)

Called when a new client connects.

fn on_client_disconnected( &mut self, client_id: usize, network: &mut NetworkServer, )

Called when a client disconnects.

fn on_message_received( &mut self, client_id: usize, msg: String, network: &mut NetworkServer, )

Called when a complete message (terminated by \n) is received.

fn on_tick(&mut self, network: &mut NetworkServer) -> bool

Called regularly on each tick of the event loop. Returns true to stop the server.

fn get_next_timeout(&self) -> Duration

Returns the duration until the next game event timeout.